To address mounting more then one filesystem i same play, i needed to change several parts of the role.
Added function to remote mount more then one filesystems in same play.
Added Cleanup of Remote Mount without restapi on client cluster.
Added more parameters when mounting filesystems. like automount, priority,
Check's and add remote filesystems if not already there.
Check if remote cluster is already defined, and skip the task.
Removed the need for JQ, uses SED and TR insted when running against no GUI on remote cluster.
Corrected the forcerun for with and without GUI on remote cluster
Changed the delay and retry to 3sec to shorten down the time of the play.
Cleaned up the naming of tasks and output from tasks.
added run_once on all of the tasks, so i someone run this against host list, it will run fine.
Added variables for validate_certs: in the URI, so we can change this if certificate are valid.
Added support to encrypt the traffic between Storage and Client Cluster with: remotecluster_chipers:
Updated the readme with new variabled and more information about network ports and troubleshooting.
Most likely more...

Limitation for now.
- Can't remove singel filesystems without removing the remote cluster connection
- Can't update filesystems Variables. like mountpoint, mount_attributes.

Signed-off-by: Ole Kristian <ole.kr.myklebust@gmail.com>